Move `RobinhoodStockTradeMapper` to a separate `stocks` folder
